🚚 사용할 수 없는 링크를 안내하는 코드를 common 디렉토리의 모듈로 분리
링크 컴포넌트 뿐만 아니라 훅도 사용해야하므로 common 디렉토리의 모듈로
분리합니다.
